+++ scarletio/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.0.59'
+__version__ = '1.0.60'
 
 from .core import *
 from .ext import *