+++ changelog.md
@@ -1,8 +1,15 @@
+## 1.0.60 *\[2023-08-10\]*
+
+#### Improvements
+
+##### ext.asyncio
+- Implement `Condition`.
+
 ## 1.0.59 *\[2023-07-28\]*
 
 #### Bug fixes
 
-- It could happen that the interactive console didnt have builtins(?).
+- It could happen that the interactive console did not have builtins(?).
 
 ## 1.0.58 *\[2023-07-28\]*